Organize PWM LiDAR and camera modules
* Move PWM and LiDAR to modules folder
* Separate camera functions from input component
